requirements.txt
DateTime==4.3
discord-webhook==0.13.0
Django==3.2.4
django-auditlog==1.0a1
django-cors-headers==3.7.0
django-jsonfield==1.4.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
python-dotenv==0.15.0
requests==2.25.1
sentry-sdk==1.3.1